requirements.txt
pyautogen>=0.2.0
openai>=1.3.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.4.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
tiktoken>=0.5.1
chromadb>=0.4.18
langchain>=0.0.335
langchain-openai>=0.0.2
